+++ include/gpu/gpu_quad.h
@@ -3,11 +3,11 @@
 #ifndef _GPU_QUAD_H_
 #define _GPU_QUAD_H_
 
-  int GPU_quad_create(void);
-  void GPU_quad_delete(const int id);
-  void GPU_quad_draw(const int id);
-  void GPU_quad_set_texture(const int id, const int tex_id);
-  void GPU_quad_set_shader(const int id, const int shader_id);
-  void GPU_quad_set_alimit(const int id, const float val);
+  short GPU_quad_create(void);
+  void GPU_quad_delete(const short id);
+  void GPU_quad_draw(const short id);
+  void GPU_quad_set_texture(const short id, const short tex_id);
+  void GPU_quad_set_shader(const short id, const short shader_id);
+  void GPU_quad_set_alimit(const short id, const float val);
 
 #endif
